+++ cedar-policy/src/api.rs
@@ -38,6 +38,7 @@ use cedar_policy_core::parser;
 pub use cedar_policy_core::parser::err::ParseErrors;
 use cedar_policy_core::parser::SourceInfo;
 use cedar_policy_core::FromNormalizedStr;
+use cedar_policy_validator::RequestValidationError; // this type is unsuitable for `pub use` because it contains internal types like `EntityUID` and `EntityType`
 pub use cedar_policy_validator::{
     TypeErrorKind, UnsupportedFeature, ValidationErrorKind, ValidationWarningKind,
 };
@@ -619,7 +620,7 @@ impl Authorizer {
     /// #
     /// # let c = Context::empty();
     /// #
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
     /// #
     /// # // create a policy
     /// # let s = r#"permit(
@@ -675,7 +676,7 @@ impl Authorizer {
     ///
     /// let c = Context::empty();
     ///
-    /// let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    /// let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
     ///
     /// // create a policy
     /// let s = r#"
@@ -801,7 +802,7 @@ impl Diagnostics {
     /// #
     /// # let c = Context::empty();
     /// #
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
     /// #
     /// # // create a policy
     /// # let s = r#"permit(
@@ -860,7 +861,7 @@ impl Diagnostics {
     /// #
     /// # let c = Context::empty();
     /// #
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
     /// #
     /// # // create a policy
     /// # let s = r#"permit(
@@ -1346,6 +1347,7 @@ impl<'a> From<cedar_policy_validator::ValidationResult<'a>> for ValidationResult
 /// and provides details specific to that kind of problem. The error also records
 /// where the problem was encountered.
 #[derive(Debug, Error)]
+#[error("validation error on `{}`: {}", self.location, self.error_kind())]
 pub struct ValidationError<'a> {
     location: SourceLocation<'a>,
     error_kind: ValidationErrorKind,
@@ -1370,22 +1372,6 @@ impl<'a> From<cedar_policy_validator::ValidationError<'a>> for ValidationError<'
             location: SourceLocation::from(location),
             error_kind,
         }
-    }
-}
-
-impl<'a> std::fmt::Display for ValidationError<'a> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "validation error on policy `{}`",
-            self.location.policy_id
-        )?;
-        if let (Some(range_start), Some(range_end)) =
-            (self.location().range_start(), self.location().range_end())
-        {
-            write!(f, " at offset {range_start}-{range_end}")?;
-        }
-        write!(f, ": {}", self.error_kind())
     }
 }
 
@@ -1415,6 +1401,21 @@ impl<'a> SourceLocation<'a> {
     }
 }
 
+impl<'a> std::fmt::Display for SourceLocation<'a> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "policy `{}`", self.policy_id)?;
+        if let Some(source_range) = &self.source_range {
+            write!(
+                f,
+                " at offset {}-{}",
+                source_range.range_start(),
+                source_range.range_end()
+            )?;
+        }
+        Ok(())
+    }
+}
+
 impl<'a> From<cedar_policy_validator::SourceLocation<'a>> for SourceLocation<'a> {
     fn from(loc: cedar_policy_validator::SourceLocation<'a>) -> SourceLocation<'a> {
         let policy_id: &'a PolicyId = PolicyId::ref_cast(loc.policy_id());
@@ -1438,7 +1439,7 @@ pub fn confusable_string_checker<'a>(
 }
 
 #[derive(Debug, Error)]
-#[error("validation warning on policy `{}`: {}", .location.policy_id, .kind)]
+#[error("validation warning on `{}`: {}", .location, .kind)]
 /// Warnings found in Cedar policies
 pub struct ValidationWarning<'a> {
     location: SourceLocation<'a>,
@@ -1462,10 +1463,7 @@ impl<'a> From<cedar_policy_validator::ValidationWarning<'a>> for ValidationWarni
     fn from(w: cedar_policy_validator::ValidationWarning<'a>) -> Self {
         let (loc, kind) = w.to_kind_and_location();
         ValidationWarning {
-            location: SourceLocation {
-                policy_id: PolicyId::ref_cast(loc),
-                source_range: None,
-            },
+            location: loc.into(),
             kind,
         }
     }
@@ -2875,28 +2873,31 @@ impl FromStr for RestrictedExpression {
 /// for partial evaluation.
 #[cfg(feature = "partial-eval")]
 #[derive(Debug)]
-pub struct RequestBuilder {
+pub struct RequestBuilder<'a> {
     principal: ast::EntityUIDEntry,
     action: ast::EntityUIDEntry,
     resource: ast::EntityUIDEntry,
     /// Here, `None` means unknown
     context: Option<ast::Context>,
+    /// Here, `None` means no request validation is performed
+    schema: Option<&'a Schema>,
 }
 
 #[cfg(feature = "partial-eval")]
-impl Default for RequestBuilder {
+impl<'a> Default for RequestBuilder<'a> {
     fn default() -> Self {
         Self {
             principal: ast::EntityUIDEntry::Unknown,
             action: ast::EntityUIDEntry::Unknown,
             resource: ast::EntityUIDEntry::Unknown,
             context: None,
+            schema: None,
         }
     }
 }
 
 #[cfg(feature = "partial-eval")]
-impl RequestBuilder {
+impl<'a> RequestBuilder<'a> {
     /// Set the principal.
     ///
     /// Note that you can create the `EntityUid` using `.parse()` on any
@@ -2968,14 +2969,24 @@ impl RequestBuilder {
         }
     }
 
+    /// Set the schema. If present, this will be used for request validation.
+    pub fn schema(self, schema: &'a Schema) -> Self {
+        Self {
+            schema: Some(schema),
+            ..self
+        }
+    }
+
     /// Create the [`Request`]
-    pub fn build(self) -> Request {
-        Request(ast::Request::new_with_unknowns(
+    pub fn build(self) -> Result<Request, RequestValidationError> {
+        Ok(Request(ast::Request::new_with_unknowns(
             self.principal,
             self.action,
             self.resource,
             self.context,
-        ))
+            self.schema.map(|schema| &schema.0),
+            Extensions::all_available(),
+        )?))
     }
 }
 
@@ -2987,7 +2998,7 @@ pub struct Request(pub(crate) ast::Request);
 impl Request {
     /// Create a [`RequestBuilder`]
     #[cfg(feature = "partial-eval")]
-    pub fn builder() -> RequestBuilder {
+    pub fn builder<'a>() -> RequestBuilder<'a> {
         RequestBuilder::default()
     }
 
@@ -3000,12 +3011,16 @@ impl Request {
     /// decisions (e.g., because they are not used in your policies).
     /// If any of the fields are `None`, we will automatically generate
     /// a unique entity UID that is not equal to any UID in the store.
+    ///
+    /// If `schema` is present, this constructor will validate that the
+    /// `Request` complies with the given `schema`.
     pub fn new(
         principal: Option<EntityUid>,
         action: Option<EntityUid>,
         resource: Option<EntityUid>,
         context: Context,
-    ) -> Self {
+        schema: Option<&Schema>,
+    ) -> Result<Self, RequestValidationError> {
         let p = match principal {
             Some(p) => p.0,
             None => ast::EntityUID::unspecified_from_eid(ast::Eid::new("principal")),
@@ -3018,7 +3033,14 @@ impl Request {
             Some(r) => r.0,
             None => ast::EntityUID::unspecified_from_eid(ast::Eid::new("resource")),
         };
-        Self(ast::Request::new(p, a, r, context.0))
+        Ok(Self(ast::Request::new(
+            p,
+            a,
+            r,
+            context.0,
+            schema.map(|schema| &schema.0),
+            Extensions::all_available(),
+        )?))
     }
 
     /// Get the principal component of the request. Returns `None` if the principal is
@@ -3073,8 +3095,7 @@ impl Context {
     /// Create an empty `Context`
     /// ```
     /// use cedar_policy::Context;
-    /// let c = Context::empty();
-    /// // let request: Request = Request::new(Some(principal), Some(action), Some(resource), c);
+    /// let context = Context::empty();
     /// ```
     pub fn empty() -> Self {
         Self(ast::Context::empty())
@@ -3085,7 +3106,7 @@ impl Context {
     /// of `(key, restricted expression)` pairs.
     /// ```
     /// use cedar_policy::{Context, RestrictedExpression};
-    /// use std::collections::HashMap;
+    /// # use std::collections::HashMap;
     /// use std::str::FromStr;
     /// # use cedar_policy::{Entities, EntityId, EntityTypeName, EntityUid, Request,PolicySet};
     /// let data : serde_json::Value = serde_json::json!({
@@ -3112,7 +3133,7 @@ impl Context {
     /// # let r_eid = EntityId::from_str("trip").unwrap();
     /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
     /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
-    /// let request: Request = Request::new(Some(p), Some(a), Some(r), context);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), context, None).unwrap();
     /// ```
     pub fn from_pairs(
         pairs: impl IntoIterator<Item = (String, RestrictedExpression)>,
@@ -3134,7 +3155,7 @@ impl Context {
     /// must specify the `Action` for schema-based parsing.
     /// ```
     /// use cedar_policy::{Context, RestrictedExpression};
-    /// use std::collections::HashMap;
+    /// # use std::collections::HashMap;
     /// use std::str::FromStr;
     /// # use cedar_policy::{Entities, EntityId, EntityTypeName, EntityUid, Request,PolicySet};
     /// let data =r#"{
@@ -3158,7 +3179,7 @@ impl Context {
     /// # let r_eid = EntityId::from_str("trip").unwrap();
     /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
     /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
-    /// let request: Request = Request::new(Some(p), Some(a), Some(r), context);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), context, None).unwrap();
     /// ```
     pub fn from_json_str(
         json: &str,
@@ -3185,33 +3206,36 @@ impl Context {
     /// must specify the `Action` for schema-based parsing.
     /// ```
     /// use cedar_policy::{Context, RestrictedExpression, Schema};
-    /// use std::collections::HashMap;
+    /// # use std::collections::HashMap;
     /// use std::str::FromStr;
     /// # use cedar_policy::{Entities, EntityId, EntityTypeName, EntityUid, Request,PolicySet};
     /// let data = serde_json::json!(
     /// {
-    ///     "sub": "1234"
+    ///     "sub": 1234
     /// });
-    /// let schema_data =r#"
+    /// let schema_json = serde_json::json!(
     ///     {
     ///       "": {
-    ///         "entityTypes": {},
-    ///           "actions": {
-    ///             "view": {
-    ///                "appliesTo": {
-    ///                  "principalTypes": [],
-    ///                   "resourceTypes": [],
-    ///                   "context": {
-    ///                     "type": "Record",
-    ///                     "attributes": {
-    ///                       "sub": { "type": "Long" }
-    ///                     }
+    ///         "entityTypes": {
+    ///           "User": {},
+    ///           "Album": {},
+    ///         },
+    ///         "actions": {
+    ///           "view": {
+    ///              "appliesTo": {
+    ///                "principalTypes": ["User"],
+    ///                 "resourceTypes": ["Album"],
+    ///                 "context": {
+    ///                   "type": "Record",
+    ///                   "attributes": {
+    ///                     "sub": { "type": "Long" }
     ///                   }
     ///                 }
     ///               }
+    ///             }
     ///           }
     ///       }
-    ///     }"#;
+    ///     });
     /// # // create a request
     /// # let p_eid = EntityId::from_str("alice").unwrap();
     /// # let p_name: EntityTypeName = EntityTypeName::from_str("User").unwrap();
@@ -3222,9 +3246,9 @@ impl Context {
     /// # let r_eid = EntityId::from_str("trip").unwrap();
     /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
     /// # let resource = EntityUid::from_type_name_and_id(r_name, r_eid);
-    /// let schema = Schema::from_str(schema_data).unwrap();
+    /// let schema = Schema::from_json_value(schema_json).unwrap();
     /// let context = Context::from_json_value(data, Some((&schema, &action))).unwrap();
-    /// let request: Request = Request::new(Some(principal), Some(action), Some(resource), context);
+    /// # let request: Request = Request::new(Some(principal), Some(action), Some(resource), context, Some(&schema)).unwrap();
     /// ```
     pub fn from_json_value(
         json: serde_json::Value,
@@ -3255,7 +3279,7 @@ impl Context {
     /// # use std::collections::HashMap;
     /// # use std::str::FromStr;
     /// # use std::fs::File;
-    /// let mut json = File::open("json_file.txt").expect("failed");
+    /// let mut json = File::open("json_file.txt").unwrap();
     /// let context = Context::from_json_file(&json, None).unwrap();
     /// # // create a request
     /// # let p_eid = EntityId::from_str("alice").unwrap();
@@ -3268,7 +3292,7 @@ impl Context {
     /// # let r_eid = EntityId::from_str("trip").unwrap();
     /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
     /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
-    /// let request: Request = Request::new(Some(p), Some(a), Some(r), context);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), context, None).unwrap();
     /// ```
     pub fn from_json_file(
         json: impl std::io::Read,
@@ -3288,12 +3312,11 @@ impl Context {
         schema: &Schema,
         action: &EntityUid,
     ) -> Result<impl ContextSchema, ContextJsonError> {
-        schema
-            .0
-            .get_context_schema(&action.0)
-            .ok_or_else(|| ContextJsonError::MissingAction {
+        cedar_policy_validator::context_schema_for_action(&schema.0, &action.0).ok_or_else(|| {
+            ContextJsonError::MissingAction {
                 action: action.clone(),
-            })
+            }
+        })
     }
 }
 
@@ -3492,6 +3515,7 @@ mod partial_eval_test {
 #[cfg(test)]
 mod entity_uid_tests {
     use super::*;
+    use cool_asserts::assert_matches;
 
     /// building an `EntityUid` from components
     #[test]
@@ -3703,13 +3727,13 @@ permit(principal ==  A :: B
     #[test]
     fn accessing_unspecified_entity_returns_none() {
         let c = Context::empty();
-        let request: Request = Request::new(None, None, None, c);
+        let request = Request::new(None, None, None, c, None).unwrap();
         let p = request.principal();
         let a = request.action();
         let r = request.resource();
-        assert!(p.is_none());
-        assert!(a.is_none());
-        assert!(r.is_none());
+        assert_matches!(p, None);
+        assert_matches!(a, None);
+        assert_matches!(r, None);
     }
 }
 
@@ -4004,7 +4028,9 @@ mod policy_set_tests {
             Some(EntityUid::from_strs("Action", "a")),
             Some(EntityUid::from_strs("Resource", "b")),
             Context::empty(),
-        );
+            None,
+        )
+        .unwrap();
 
         let e = r#"[
             {
@@ -4286,7 +4312,9 @@ mod policy_set_tests {
             Some(EntityUid::from_strs("Action", "a")),
             Some(EntityUid::from_strs("Resource", "b")),
             Context::empty(),
-        );
+            None,
+        )
+        .unwrap();
 
         let e = r#"[
             {