+++ cedar-policy-cli/CHANGELOG.md
@@ -6,6 +6,16 @@
 
 - `--deny-warnings` option to `validate` command. This option turns non-fatal
   warnings into errors.
+- Requests are now validated by default if a schema is provided. This can be
+  disabled with `--request-validation=false`.
+- The `-s` short form can now be used for `--schema` across all subcommands.
+
+### Changed
+
+- The `-p` flag now always refers to `--policies` (not `--principal`) across all
+  subcommands, while `-l` refers to `--principal`. Relatedly, the `--policies`
+  long form of the flag is also now accepted across all subcommands.
+- The short form of `--template-linked` was changed from `-t` to `-k`.
 
 ## 2.4.2